% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11altrepbug-package.R
\docType{package}
\name{cpp11altrepbug-package}
\alias{cpp11altrepbug}
\alias{cpp11altrepbug-package}
\title{cpp11altrepbug: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
